onboarding-checklist/backend/requirements.txt
alembic==1.10.3
async-timeout==4.0.2
bcrypt==4.0.1
cachelib==0.10.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
docopt==0.6.2
flash==1.0.3
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Sessionstore==0.4.5
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
numpy==1.23.4
pandas==1.5.3
pipreqs==0.4.13
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.6
redis==4.5.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarg==0.1.9